GEODE-4268: Delete geode-benchmarks and move benchmarks to geode-core
