+++ geode-core/src/jmh/java/org/apache/geode/cache/EvictionBasePerformanceBenchmark.java
@@ -12,7 +12,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package org.apache.geode.cache.benchmark;
+package org.apache.geode.cache;
 
 import static org.apache.geode.distributed.ConfigurationProperties.LOG_LEVEL;
 
@@ -26,32 +26,31 @@ import org.openjdk.jmh.annotations.Level;
 import org.openjdk.jmh.annotations.Measurement;
 import org.openjdk.jmh.annotations.Mode;
 import org.openjdk.jmh.annotations.OutputTimeUnit;
+import org.openjdk.jmh.annotations.Param;
 import org.openjdk.jmh.annotations.Scope;
 import org.openjdk.jmh.annotations.Setup;
 import org.openjdk.jmh.annotations.State;
 import org.openjdk.jmh.annotations.TearDown;
 import org.openjdk.jmh.annotations.Warmup;
 
-import org.apache.geode.cache.Cache;
-import org.apache.geode.cache.CacheFactory;
-import org.apache.geode.cache.Region;
-import org.apache.geode.cache.RegionShortcut;
+import org.apache.geode.internal.lang.SystemPropertyHelper;
 
-/**
- * This benchmark measures the raw throughput of get actions on a region
- */
 @State(Scope.Thread)
 @Fork(1)
-public class GetOnRegionBenchmark {
-  private static final int ENTRIES = 1_000_000;
+public class EvictionBasePerformanceBenchmark {
+  private static final int MAX_ENTRIES = 1_000_000;
+
+  @Param({"true", "false"})
+  public String useAsync;
 
   Cache cache;
   Region<String, String> region;
 
   @Setup(Level.Trial)
   public void setup() {
+    System.setProperty("geode." + SystemPropertyHelper.EVICTION_SCAN_ASYNC, useAsync);
     cache = new CacheFactory().set(LOG_LEVEL, "warn").create();
-    region = createRegion(cache);
+    region = createRegion(cache, MAX_ENTRIES);
   }
 
   @TearDown(Level.Trial)
@@ -62,6 +61,7 @@ public class GetOnRegionBenchmark {
   @State(Scope.Thread)
   public static class MyState {
     Random random = new Random();
+    int nextKey = MAX_ENTRIES + 1;
   }
 
   @Benchmark
@@ -69,17 +69,30 @@ public class GetOnRegionBenchmark {
   @Warmup(iterations = 5)
   @BenchmarkMode(Mode.Throughput)
   @OutputTimeUnit(TimeUnit.MILLISECONDS)
-  public String createEntry(MyState state) {
-    String key = Integer.toString(state.random.nextInt(ENTRIES));
-    return region.get(key);
+  public String nonEvictingUpdate(MyState state) {
+    String key = Integer.toString(state.random.nextInt(MAX_ENTRIES));
+    return region.put(key, "value");
   }
 
-  private Region<String, String> createRegion(Cache cache) {
-    Region<String, String> region =
-        cache.<String, String>createRegionFactory(RegionShortcut.LOCAL).create("testRegion");
-    for (int i = 0; i < ENTRIES; i++) {
+  @Benchmark
+  @Measurement(iterations = 50)
+  @Warmup(iterations = 5)
+  @BenchmarkMode(Mode.Throughput)
+  @OutputTimeUnit(TimeUnit.MILLISECONDS)
+  public String evictingCreate(MyState state) {
+    String key = Integer.toString(state.nextKey++);
+    return region.put(key, "value");
+  }
+
+  private Region<String, String> createRegion(Cache cache, int maxSize) {
+    Region<String, String> region = cache.<String, String>createRegionFactory(RegionShortcut.LOCAL)
+        .setEvictionAttributes(
+            EvictionAttributes.createLRUEntryAttributes(maxSize, EvictionAction.LOCAL_DESTROY))
+        .create("testRegion");
+    for (int i = 0; i < MAX_ENTRIES; i++) {
       region.put(Integer.toString(i), "value");
     }
+    region.put("over", "limit");
     return region;
   }
 }
