+++ settings.gradle
@@ -31,7 +31,6 @@ include 'geode-old-client-support'
 include 'geode-wan'
 include 'geode-cq'
 include 'geode-connectors'
-include 'geode-benchmarks'
 include 'geode-client-protocol'
 include 'extensions/geode-modules'
 include 'extensions/geode-modules-tomcat7'
